requirements.txt
# HTTP Client
requests>=2.31.0,<3.0.0

# Database
supabase>=2.22.0,<3.0.0
httpx>=0.28.0,<1.0.0
websockets>=15.0.0,<16.0.0

# AI / OpenAI (Phase 7)
openai>=1.57.0,<2.0.0

# Environment
python-dotenv>=1.0.0,<2.0.0
